JWS-1151 Add database testing support
Please refer to DataSourceFactory javadoc as an entry point how to use
the class for creating a pooling data source without tomcat.
